fastapi-blog/requirements.txt
fastapi
uvicorn[standard]
asyncpg
sqlalchemy
psycopg2-binary
python-dotenv
alembic
pydantic[email] 
pytest
httpx
python-jose[cryptography]
passlib[bcrypt]
bcrypt==3.2.2
python-multipart
pydantic-settings
cloudinary
celery
redis
aiosmtplib
uvicorn
websockets